deleted some files, created some AppComponents
