+++ src/app/store.ts
@@ -1,10 +1,8 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
 import userReducer from "./slices/user";
 
 export const store = configureStore({
   reducer: {
-    counter: counterReducer,
     user: userReducer
   },
 });
