--- src/components/AppComponents/AppComponent.module.scss
@@ -1,71 +1,0 @@
-.container {
-  margin: 0 auto;
-  width: 100%;
-  max-width: 470px;
-}
-
-.avatar {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  border: 1px solid red;
-  overflow: hidden;
-  > img {
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-  }
-}
-
-.button {
-  background-color: #0095f6;
-  color: #fff;
-  font-weight: 500;
-  outline: none;
-  border: none;
-  border-radius: 8px;
-  padding: 7px 16px;
-  transition: .2s;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #1877f2;
-  }
-}
-
-.disabled {
-  opacity: 0.5;
-  pointer-events: none;
-}
-
-.gray {
-  background-color: #efefef;
-  color: #000;
-
-  &:hover {
-    background-color: #bdbdbd;
-  }
-}
-
-.loading__page {
-  width: 100vw;
-  height: 100vh;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  >svg {
-    width: 70px;
-    color: #f40866;
-  }
-
-  >p {
-    font-size: 20px;
-    font-weight: 500;
-    background: linear-gradient(135deg, #920af1 0%, #f71931 59%, #f40866 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-  }
-}
