+++ src/components/AppComponents/Avatar/Avatar.module.scss
@@ -1,0 +1,13 @@
+.avatar {
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  border: 1px solid red;
+  overflow: hidden;
+
+  >img {
+    object-fit: cover;
+    width: 100%;
+    height: 100%;
+  }
+}
