+++ src/components/RouteComponents/PrivateRoute.tsx
@@ -1,8 +1,8 @@
 import { Navigate, Outlet, useLocation } from "react-router-dom";
-import { useAppSelector } from "../app/hooks";
-import { selectUser } from "../app/slices/user";
-import { LoadingPage } from "./AppComponents";
-import { SideBar } from "./SideBar";
+import { useAppSelector } from "../../app/hooks";
+import { selectUser } from "../../app/slices/user";
+import { LoadingPage } from "../../pages";
+import { SideBar } from "../SideBar";
 
 export const PrivateRoute: React.FC = () => {
   const { user, status } = useAppSelector(selectUser);
@@ -15,9 +15,11 @@ export const PrivateRoute: React.FC = () => {
   if (!user) {
     return <Navigate to="/auth/login" state={location.pathname} />;
   }
-	
-  return <>
-		<SideBar />
-		<Outlet />
-	</>;
+
+  return (
+    <>
+      <SideBar />
+      <Outlet />
+    </>
+  );
 };
