+++ src/components/modals/modals.module.scss
@@ -92,81 +92,6 @@
   }
 }
 
-.upload__title {
-  padding: 9px 16px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  >p:only-child {
-    width: 100%;
-  }
-
-  >div {
-    width: 24px;
-    height: 24px;
-    cursor: pointer;
-  }
-
-  >.next {
-    color: #0095f6;
-    transition: .2s;
-
-    &:hover {
-      color: #00376b;
-    }
-  }
-}
-
-.upload__inner {
-  position: relative;
-  width: 680px;
-  height: 680px;
-  // width: auto;
-  // height: auto;
-  padding: 24px;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  overflow: hidden;
-
-  input[type=file] {
-    display: none;
-  }
-
-  >.arrow {
-    position: absolute;
-    z-index: 100;
-  }
-
-  >.left {
-    left: 4%;
-  }
-
-  >.right {
-    right: 4%;
-  }
-
-  >.loadingSvg {
-    >svg {
-      color: #0095f6;
-      width: 56px;
-    }
-  }
-
-  >svg {
-    width: 97px;
-  }
-
-  >p {
-    font-size: 20px;
-    margin: 16px 0;
-  }
-}
-
 .filter>ul {
   display: flex;
 
@@ -248,382 +173,6 @@
         color: #ed4956;
         font-weight: 500;
       }
-    }
-  }
-}
-
-.upload__images {
-  list-style: none;
-  position: absolute;
-  left: 0;
-  height: 100%;
-  width: 100%;
-  display: flex;
-  transition: .2s;
-  user-select: none;
-
-  >li {
-    overflow: hidden;
-    height: 100%;
-    width: 100%;
-    min-width: 680px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 0 0 12px 12px;
-    position: relative;
-
-    >img {
-      position: absolute;
-      // top: 50%;
-      // left: 50%;
-      // min-height: 100%;
-      // min-width: 100%;
-      cursor: grab;
-      // transform: translate(-50%, -50%);
-      // object-fit: cover;
-      // width: 100%;
-      // height: 100%;
-    }
-  }
-}
-
-.upload__images__small {
-  position: absolute;
-  z-index: 100;
-  bottom: 64px;
-  right: 4%;
-  padding: 8px;
-  border-radius: 8px;
-  max-width: 92%;
-  background-color: rgba(26, 26, 26, .8);
-  display: flex;
-
-  >.images {
-    position: relative;
-    overflow: hidden;
-
-
-    >.arrow {
-      position: absolute;
-      z-index: 110;
-      top: 50%;
-      transform: translateY(-50%);
-      width: 26px;
-      height: 26px;
-
-      >.icon__button {
-        width: 26px;
-        height: 26px;
-        background-color: #fff;
-
-        >svg {
-          color: #000;
-        }
-      }
-    }
-
-    >.right {
-      right: 10px;
-    }
-
-    >.left {
-      left: 10px;
-    }
-
-    >ul {
-      display: flex;
-      overflow-x: scroll;
-      position: relative;
-
-      // >.arrow {
-      //   position: absolute;
-      //   z-index: 110;
-      //   top: 50%;
-      //   transform: translateY(-50%);
-      //   width: 26px;
-      //   height: 26px;
-
-      //   >div {
-      //     width: 26px;
-      //     height: 26px;
-      //     background-color: #fff;
-      //     position: fixed;
-
-      //     >svg {
-      //       color: #000;
-      //     }
-      //   }
-
-      // }
-
-      // >.right {
-      //   right: 0;
-      // }
-
-      >.active {
-        opacity: 1;
-
-        >div {
-          display: block;
-        }
-      }
-
-      >li {
-        position: relative;
-        overflow: hidden;
-        width: 94px !important;
-        height: 94px;
-        margin: 6px;
-        flex: none;
-        cursor: pointer;
-        opacity: .7;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-
-        >div {
-          display: none;
-          position: absolute;
-          z-index: 200;
-          width: 20px;
-          height: 20px;
-          border-radius: 50%;
-          background-color: rgba(26, 26, 26, .8);
-          top: 8px;
-          right: 8px;
-          transition: .2s;
-          cursor: pointer;
-
-          &:hover {
-            opacity: .7;
-          }
-
-          >svg {
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-            color: #fff;
-            width: 12px;
-            height: 12px;
-          }
-        }
-
-        >img {
-          position: absolute;
-          // min-height: 100%;
-          // min-width: 100%;
-          // top: 50%;
-          // left: 50%;
-          z-index: 110;
-          // transform: scale(calc(94 / 680));
-          // object-fit: cover;
-          // width: 100%;
-          // height: 100%;
-        }
-      }
-
-      &::-webkit-scrollbar {
-        display: none;
-      }
-    }
-  }
-
-  >.plus {
-    flex: none;
-    width: 48px;
-    height: 48px;
-    border-radius: 50%;
-    margin-left: 6px;
-    border: 1px solid #dbdbdb;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-    transition: .2s;
-
-    &:active {
-      background-color: #efefef;
-    }
-
-    >svg {
-      fill: #8e8e8e;
-    }
-  }
-
-
-
-}
-
-// .arrows {
-//   position: absolute;
-//   z-index: 100;
-//   width: 95%;
-//   height: 32px;
-
-//   >.left {
-//     left: 0;
-//   }
-
-//   >.right {
-//     right: 0;
-//   }
-
-//   // &:only-child {
-//   //   justify-content: flex-end;
-//   // }
-
-//   >div {
-//     position: absolute;
-//   }
-// }
-
-
-.upload__sub {
-  position: absolute;
-  bottom: 24px;
-  z-index: 100;
-  width: 92%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  .active {
-    background-color: #fff;
-
-    >svg {
-      fill: #262626;
-    }
-  }
-
-  .disabled {
-    opacity: .3 !important;
-  }
-
-  >div:first-child {
-    display: flex;
-
-    >div:first-child {
-      margin-right: 8px;
-    }
-  }
-}
-
-.icon__button {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  background-color: rgba(26, 26, 26, .8);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  transition: .2s;
-
-  &:hover {
-    opacity: .7;
-  }
-
-  >svg {
-    width: 12px;
-    height: 12px;
-    fill: #fff;
-    color: #fff;
-  }
-}
-
-.dots {
-  list-style: none;
-  display: flex;
-
-  >li {
-    width: 6px;
-    height: 6px;
-    margin: 0 2px;
-    border-radius: 50%;
-    background-color: #fff;
-    opacity: 0.4;
-    transition: .2s;
-    cursor: pointer;
-  }
-}
-
-.mouseMove {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  z-index: 10000;
-}
-
-.cropMenu {
-  background-color: rgba(26, 26, 26, .8);
-  border-radius: 8px;
-  position: absolute;
-  left: 4%;
-  bottom: 64px;
-
-  >li {
-    border-bottom: 1px solid #dbdbdb;
-    color: #8e8e8e;
-    padding-left: 12px;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-
-    &:last-child {
-      border-bottom: none;
-    }
-
-    >svg {
-      margin: 12px;
-      color: #8e8e8e;
-      width: 24px;
-    }
-  }
-
-  >.active {
-    color: #fff;
-
-    >svg {
-      color: #fff;
-    }
-  }
-}
-
-.zoom {
-  position: absolute;
-
-  >input {
-    
-    height: 1px;
-    // appearance: none;
-    background-color: #000;
-
-    &::-moz-range-track {
-      appearance: none;
-      -moz-appearance: none;
-      background-color: #fff;
-    }
-
-    &::-webkit-slider-thumb {
-      -webkit-appearance: none;
-      height: 20px;
-      width: 20px;
-      // border-radius: 50%;
-      background: #ff4500;
-      cursor: ew-resize;
-      // box-shadow: 0 0 2px 0 #555;
-      transition: background .3s ease-in-out;
-      // -webkit-appearance: none;
-      // appearance: none;
-      // background-color: #fff;
-      // background-color: #fff;
-    }
-
-    &::-webkit-slider-runnable-track {
-      // background-color: #000;
-      // height: 1px;
     }
   }
 }
