+++ src/pages/Home/Home.module.scss
@@ -1,0 +1,5 @@
+.container {
+  margin: 0 auto;
+  width: 100%;
+  max-width: 470px;
+}
