+++ src/pages/Home/index.tsx
@@ -1,17 +1,15 @@
-import { AppContainer, Post } from "../../components";
+import { Post } from "../../components";
 import scss from "./Home.module.scss";
 
 export const Home: React.FC = () => {
   return (
     <>
-      <AppContainer>
-        <ul>
-          <Post />
-          <Post />
-          <Post />
-          <Post />
-        </ul>
-      </AppContainer>
+      <ul className={scss.container}>
+        <Post />
+        <Post />
+        <Post />
+        <Post />
+      </ul>
     </>
   );
 };
