+++ src/pages/LoadingPage/index.tsx
@@ -1,0 +1,11 @@
+import { LogoIcon } from "../../components/icons";
+import scss from "./LoadingPage.module.scss";
+
+export const LoadingPage: React.FC = () => {
+  return (
+    <div className={scss.loading__page}>
+      <LogoIcon />
+      <p>from Sam</p>
+    </div>
+  );
+};
