+++ src/pages/index.ts
@@ -3,3 +3,4 @@ export { Profile } from "./Profile";
 export { Messages } from "./Messages";
 export { Register } from "./Register";
 export { Login } from "./Login";
+export { LoadingPage } from "./LoadingPage";
